requirements.txt
vk==3.0
fastapi==0.103.2
uvicorn==0.20.0
python-dotenv==1.0.0
PyYAML==6.0.1